@@ -1,4 +1,5 @@
 import 'dart:convert';
+import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:flutter_echarts/flutter_echarts.dart';
 import '../gl_script.dart';
@@ -18,9 +19,34 @@ class Atlantic extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    if (maritimeData == null || maritimeData!.isEmpty) {
+      return const CircularProgressIndicator();
+    }
+
+    // maritimeDataから3番目の値（z軸）を抽出して最小値・最大値を計算
+    List<double> zValues = maritimeData!
+        .map((item) => item['value'][2] as double)
+        .toList();
+
+    double minZ = zValues.reduce(min).floorToDouble(); // 最小値を切り捨て
+    double maxZ = zValues.reduce(max).ceilToDouble(); // 最大値を切り上げ
+    double midZ = (minZ + maxZ) / 2; // 中間値を計算
+
+    // coastLine, ridgeLine, trenchLine の z 軸値を midZ に設定
+    List<dynamic> transformedCoastLine = coastLine?.map((coordinate) {
+      return [coordinate[0], coordinate[1], midZ];
+    }).toList() ?? [];
+
+    List<dynamic> transformedRidgeLine = ridgeLine?.map((coordinate) {
+      return [coordinate[0], coordinate[1], midZ];
+    }).toList() ?? [];
+
+    List<dynamic> transformedTrenchLine = trenchLine?.map((coordinate) {
+      return [coordinate[0], coordinate[1], midZ];
+    }).toList() ?? [];
+
     return Scaffold(
-      body: Container(
-        constraints: const BoxConstraints.expand(),
+      body: DecoratedBox(
         decoration: const BoxDecoration(
           image: DecorationImage(
             image: AssetImage('assets/images/sea.png'),
@@ -29,11 +55,9 @@ class Atlantic extends StatelessWidget {
         ),
         child: Center(
           child: SizedBox(
-            width: 1200,
-            height: 1000,
-            child: maritimeData == null
-                ? const CircularProgressIndicator()
-                : Echarts(
+            width: double.infinity,
+            height: double.infinity,
+            child: Echarts(
               extensions: const [glScript],
               option: '''
     (function() {
@@ -75,8 +99,8 @@ class Atlantic extends StatelessWidget {
         },
         zAxis3D: {
           type: 'value',
-          min: -1000,
-          max: 1000,
+          min: $minZ,
+          max: $maxZ,
           splitLine: {show: false},
           name: 'timeline',
           axisLine: {
@@ -103,23 +127,23 @@ class Atlantic extends StatelessWidget {
             {
               type: 'scatter3D',
               symbolSize: 3,
-              data: ${json.encode(coastLine)},
+              data: ${json.encode(transformedCoastLine)},
               itemStyle: { color: 'white' } 
             }, 
             {
               type: 'scatter3D',
               symbolSize: 3,
-              data: ${json.encode(ridgeLine)},
+              data: ${json.encode(transformedRidgeLine)},
                 itemStyle: {color: '#bc8f8f'}              
             }, 
             {
               type: 'scatter3D',
               symbolSize: 3,
-              data: ${json.encode(trenchLine)},
-                itemStyle: {color: '#800000'}              
+              data: ${json.encode(transformedTrenchLine)},
+                itemStyle: {color: '#cd5c5c'}              
             },       
           ]
-        };
+                  };
             })()
             ''',
             ),
